Etiquetas/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
aiomysql
greenlet
cryptography